Finishing db CRUD

+++ src/main/java/com/kuizu/exammicroservice/dao/ExamDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ExamDao extends JpaRepository<ExamEntity, Integer> {
+public interface ExamDao extends JpaRepository<ExamEntity, Long> {
     Optional<List<ExamEntity>> findAllByIdCourse(String idCourse);
 }
